+++ packages/ember-rdfa-editor/src/utils/_private/lblod-utils/insert-article.ts
@@ -1,7 +1,6 @@
 import { EditorState, TextSelection, Transaction } from 'prosemirror-state';
 import { PNode } from '#root/prosemirror-aliases.ts';
 import { ELI, PROV } from '#root/utils/_private/lblod-utils/constants.ts';
-import { getOutgoingTriple } from '#root/utils/namespace.ts';
 import {
   transactionCombinator,
   type TransactionMonad,
@@ -10,6 +9,14 @@ import { addPropertyToNode, findNodeByRdfaId } from '#root/utils/rdfa-utils.ts';
 import { SayDataFactory } from '#root/core/say-data-factory/index.ts';
 import { getNodesBySubject } from '#root/plugins/rdfa-info/index.ts';
 import { recalculateNumbers } from '#root/utils/_private/lblod-utils/recalculate-structure-numbers.ts';
+import { GapCursor } from '#root/plugins/gap-cursor/index.ts';
+import {
+  getOutgoingTriple,
+  hasOutgoingNamedNodeTriple,
+} from '#root/utils/namespace.ts';
+import { RDF } from '#root/utils/_private/lblod-utils/constants.ts';
+import { findAncestors } from '#root/utils/position-utils.ts';
+import { ResolvedPos } from 'prosemirror-model';
 
 export interface InsertArticleToDecisionArgs {
   node: PNode;
@@ -39,7 +46,9 @@ export function insertArticle(
     let replacementTr: Transaction;
     let insertLocation: number | undefined;
     if ('insertFreely' in args && args.insertFreely) {
-      replacementTr = tr.replaceSelectionWith(node);
+      // Avoid insertion inside another article
+      insertLocation = getClosestValidPosition(state.selection.$from, node);
+      replacementTr = tr.replaceWith(insertLocation, insertLocation, node);
     } else {
       const { position } = args;
       const decision = getNodesBySubject(state, args.decisionUri)[0];
@@ -89,12 +98,18 @@ export function insertArticle(
       recalculateNumbers,
     ]);
 
-    if (insertLocation) {
+    if (!args.insertFreely) {
       transaction.setSelection(
         TextSelection.create(
           transaction.doc,
           insertLocation + 1,
           insertLocation + node.nodeSize - 1,
+        ),
+      );
+    } else {
+      transaction.setSelection(
+        new GapCursor(
+          transaction.doc.resolve(transaction.mapping.map(insertLocation)),
         ),
       );
     }
@@ -140,4 +155,22 @@ function resolveInsertLocation(
   }
 
   return containerLocation.pos + containerNode.nodeSize - 1;
+}
+
+function getClosestValidPosition(position: ResolvedPos, node: PNode) {
+  const parentArticle = findAncestors(
+    position,
+    isSameArticleNode.bind(undefined, node),
+  )[0];
+  if (parentArticle) {
+    return parentArticle.pos + parentArticle.node.nodeSize;
+  } else {
+    return position.pos;
+  }
+}
+
+export function isSameArticleNode(nodeA: PNode, nodeB: PNode) {
+  const rdfType = getOutgoingTriple(nodeA.attrs, RDF('type'))?.object.value;
+  if (!rdfType) return false;
+  return hasOutgoingNamedNodeTriple(nodeB.attrs, RDF('type'), rdfType);
 }