+++ packages/ember-rdfa-editor/src/utils/_private/rdfa-parser/rdfa-parser.ts
@@ -31,6 +31,7 @@ import { postProcessTagAsRdfaNode } from './post-process-as-rdfa-nodes.ts';
 import { sayDataFactory } from '#root/core/say-data-factory/index.ts';
 import { LANG_STRING } from '../constants.ts';
 import { N3StoreWrapper } from '../datastore/n3-store-wrapper.ts';
+import { defaultPrefixes } from '#root/config/rdfa.ts';
 
 export type ModelTerm<N> =
   ModelQuadObject<N> | ModelQuadPredicate<N> | ModelQuadSubject<N>;
@@ -77,6 +78,7 @@ export interface RdfaParseConfig<N> {
 
   baseIRI: string;
   pathFromDomRoot?: Node[];
+  defaultPrefixes?: Record<string, string>;
 }
 
 export interface QuadNodes<N> {
@@ -188,7 +190,7 @@ export class RdfaParser<N> {
           ? INITIAL_CONTEXT_XHTML['@context']
           : {}),
       },
-      prefixesCustom: {},
+      prefixesCustom: options.defaultPrefixes ?? defaultPrefixes,
       skipElement: false,
       vocab: options.vocab,
       node: this.rootModelNode,
@@ -197,7 +199,11 @@ export class RdfaParser<N> {
 
   static parse<N>(config: RdfaParseConfig<N>): RdfaParseResponse<N> {
     const { pathFromDomRoot = [], root, baseIRI, parseRoot = true } = config;
-    const parser = new RdfaParser<N>({ rootModelNode: root, baseIRI });
+    const parser = new RdfaParser<N>({
+      rootModelNode: root,
+      baseIRI,
+      defaultPrefixes: config.defaultPrefixes,
+    });
     for (const domNode of pathFromDomRoot) {
       if (isElement(domNode)) {
         const attributeObj: Record<string, string> = {};
@@ -1472,6 +1478,8 @@ export interface IRdfaParserOptions<N> {
   htmlParseListener?: IHtmlParseListener;
 
   rootModelNode: N;
+
+  defaultPrefixes?: Record<string, string>;
 }
 
 /**